requirements.txt
streamlit
google-generativeai
supabase-py
